fix(student-deletion): disallow super admin account deletion

+++ src/main/java/com/iseplife/api/controllers/StudentController.java
@@ -39,6 +39,7 @@ import com.iseplife.api.dto.view.MatchedView;
 import com.iseplife.api.entity.feed.Feed;
 import com.iseplife.api.entity.user.Role;
 import com.iseplife.api.entity.user.Student;
+import com.iseplife.api.exceptions.http.HttpBadRequestException;
 import com.iseplife.api.services.ClubService;
 import com.iseplife.api.services.MediaService;
 import com.iseplife.api.services.NotificationService;
@@ -142,6 +143,8 @@ public class StudentController {
   @DeleteMapping("/{id}")
   @RolesAllowed({Roles.ADMIN, Roles.USER_MANAGER})
   public void deleteStudent(@PathVariable Long id) {
+    if(id == 1)
+      throw new HttpBadRequestException("Can't delete admin user");
     studentService.deleteStudent(id);
   }
